Placing footer
Wanted to help you more, but forgot I even posted on reddit. SO here is some more, and only had a few minutes before I have to leave. 

I made changes so it fluidly adjusts to viewport size. Should even fit mobile phones, but won't look that great. Also changed the footer so it sticks to the bottom. 

Stu